igreja_dashboard_python/pyproject.toml
[project]
name = "igreja-dashboard"
version = "0.1.0"
description = "Painel de gestão social para igrejas"
requires-python = ">=3.12,<3.13"
dependencies = [
    "alembic>=1.13.0",
    "fastapi>=0.110.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.0",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.27.0",
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["B008"]

[tool.ruff.isort]
combine-as-imports = true
force-sort-within-sections = true
